test: integrate s3 tests in storage layer

+++ storages/storage.spec.js
@@ -2,9 +2,12 @@
 const { join, parse } = require('node:path');
 const { readFile, writeFile } = require('node:fs/promises');
 const { promisify } = require('node:util');
+const { Readable } = require('node:stream');
 const mkdirp = require('mkdirp');
 const rmrf = require('rimraf');
 const { expect } = require('chai');
+const { S3Client, PutObjectCommand, GetObjectCommand } = require('@aws-sdk/client-s3');
+const { mockClient } = require('aws-sdk-client-mock');
 
 const rimraf = promisify(rmrf);
 
@@ -19,19 +22,26 @@ const {
   getProvider,
 } = require('./storage');
 
-const storageTypes = ['file'];
+const storageTypes = ['file', 's3'];
 
 storageTypes.forEach((storageType) => {
   describe(`${storageType} storage`, async () => {
+    let s3Mock;
     beforeEach(async () => {
       if (storageType === 'file') {
         process.env.CITIZEN_STORAGE_PATH = '/tmp/citizen-test';
+      } else if (storageType === 's3') {
+        s3Mock = mockClient(S3Client);
       }
       await init(storageType);
     });
 
     afterEach(async () => {
-      await rimraf(process.env.CITIZEN_STORAGE_PATH);
+      if (storageType === 'file') {
+        await rimraf(process.env.CITIZEN_STORAGE_PATH);
+      } else if (storageType === 's3') {
+        s3Mock.reset();
+      }
     });
 
     it(`should use ${storageType}`, () => {
@@ -47,33 +57,51 @@ storageTypes.forEach((storageType) => {
       });
 
       describe('saveModule()', () => {
-        it('should save the module onto the storage with relative path', async () => {
-          process.env.CITIZEN_STORAGE_PATH = './tmp/citizen-test';
-          const modulePath = `${new Date().getTime()}/module.tar.gz`;
-          const result = await saveModule(modulePath, moduleBuf);
-          expect(result).to.be.true;
-        });
+        if (storageType === 'file') {
+          it('should save the module onto the storage with relative path', async () => {
+            process.env.CITIZEN_STORAGE_PATH = './tmp/citizen-test';
+            const modulePath = `${new Date().getTime()}/module.tar.gz`;
+            const result = await saveModule(modulePath, moduleBuf);
+            expect(result).to.be.true;
+          });
 
-        it('should save the module onto the storage with absolute path', async () => {
-          const modulePath = `${new Date().getTime()}/module.tar.gz`;
-          const result = await saveModule(modulePath, moduleBuf);
-          expect(result).to.be.true;
-        });
+          it('should save the module onto the storage with absolute path', async () => {
+            const modulePath = `${new Date().getTime()}/module.tar.gz`;
+            const result = await saveModule(modulePath, moduleBuf);
+            expect(result).to.be.true;
+          });
+        }
+
+        if (storageType === 's3') {
+          it('should save the module onto S3', async () => {
+            s3Mock.on(PutObjectCommand).resolves({ ETag: '1234' });
+            const modulePath = `${new Date().getTime()}/module.tar.gz`;
+            const result = await saveModule(modulePath, moduleBuf);
+            expect(result).to.be.true;
+          });
+        }
       });
 
       describe('hasModule()', () => {
         it('should return true if the module is already exist', async () => {
           const modulePath = `${new Date().getTime()}/module.tar.gz`;
-          const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'modules', modulePath);
-          const parsedPath = parse(pathToStore);
-          await mkdirp(parsedPath.dir);
-          await writeFile(pathToStore, moduleBuf);
+          if (storageType === 'file') {
+            const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'modules', modulePath);
+            const parsedPath = parse(pathToStore);
+            await mkdirp(parsedPath.dir);
+            await writeFile(pathToStore, moduleBuf);
+          } else if (storageType === 's3') {
+            s3Mock.on(GetObjectCommand).resolves({ Body: 'data' });
+          }
 
           const exist = await hasModule(modulePath);
           expect(exist).to.be.true;
         });
 
         it('should return false if the module is not already exist', async () => {
+          if (storageType === 's3') {
+            s3Mock.on(GetObjectCommand).rejects({ name: 'NoSuchKey' });
+          }
           const modulePath = `${new Date().getTime()}/module.tar.gz`;
           const exist = await hasModule(`${modulePath}/wrong`);
           expect(exist).to.be.false;
@@ -83,10 +111,17 @@ storageTypes.forEach((storageType) => {
       describe('getModule()', () => {
         it('should get file buffer from the storage', async () => {
           const modulePath = `${new Date().getTime()}/module.tar.gz`;
-          const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'modules', modulePath);
-          const parsedPath = parse(pathToStore);
-          await mkdirp(parsedPath.dir);
-          await writeFile(pathToStore, moduleBuf);
+          if (storageType === 'file') {
+            const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'modules', modulePath);
+            const parsedPath = parse(pathToStore);
+            await mkdirp(parsedPath.dir);
+            await writeFile(pathToStore, moduleBuf);
+          } else if (storageType === 's3') {
+            const buf = Buffer.from('data');
+            s3Mock.on(GetObjectCommand).resolves({
+              Body: Readable.from(buf),
+            });
+          }
 
           const result = await getModule(modulePath);
           expect(result).to.be.an.instanceof(Buffer);
@@ -110,33 +145,51 @@ storageTypes.forEach((storageType) => {
       });
 
       describe('saveProvider()', () => {
-        it('should save the provider onto the storage with relative path', async () => {
-          process.env.CITIZEN_STORAGE_PATH = './tmp/citizen-test';
-          const providerPath = `${new Date().getTime()}/provider.tar.gz`;
-          const result = await saveProvider(providerPath, providerBuf);
-          expect(result).to.be.true;
-        });
+        if (storageType === 'file') {
+          it('should save the provider onto the storage with relative path', async () => {
+            process.env.CITIZEN_STORAGE_PATH = './tmp/citizen-test';
+            const providerPath = `${new Date().getTime()}/provider.tar.gz`;
+            const result = await saveProvider(providerPath, providerBuf);
+            expect(result).to.be.true;
+          });
 
-        it('should save the provider onto the storage with absolute path', async () => {
-          const providerPath = `${new Date().getTime()}/provider.tar.gz`;
-          const result = await saveProvider(providerPath, providerBuf);
-          expect(result).to.be.true;
-        });
+          it('should save the provider onto the storage with absolute path', async () => {
+            const providerPath = `${new Date().getTime()}/provider.tar.gz`;
+            const result = await saveProvider(providerPath, providerBuf);
+            expect(result).to.be.true;
+          });
+        }
+
+        if (storageType === 's3') {
+          it('should save the provider onto S3', async () => {
+            s3Mock.on(PutObjectCommand).resolves({ ETag: '1234' });
+            const providerPath = `${new Date().getTime()}/provider.tar.gz`;
+            const result = await saveProvider(providerPath, providerBuf);
+            expect(result).to.be.true;
+          });
+        }
       });
 
       describe('hasProvider()', () => {
         it('should return true if the provider is already exist', async () => {
           const providerPath = `${new Date().getTime()}/provider.tar.gz`;
-          const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'providers', providerPath);
-          const parsedPath = parse(pathToStore);
-          await mkdirp(parsedPath.dir);
-          await writeFile(pathToStore, providerBuf);
+          if (storageType === 'file') {
+            const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'providers', providerPath);
+            const parsedPath = parse(pathToStore);
+            await mkdirp(parsedPath.dir);
+            await writeFile(pathToStore, providerBuf);
+          } else if (storageType === 's3') {
+            s3Mock.on(GetObjectCommand).resolves({ Body: 'data' });
+          }
 
           const exist = await hasProvider(providerPath);
           expect(exist).to.be.true;
         });
 
         it('should return false if the provider is not already exist', async () => {
+          if (storageType === 's3') {
+            s3Mock.on(GetObjectCommand).rejects({ name: 'NoSuchKey' });
+          }
           const providerPath = `${new Date().getTime()}/provider.tar.gz`;
           const exist = await hasProvider(`${providerPath}/wrong`);
           expect(exist).to.be.false;
@@ -146,10 +199,17 @@ storageTypes.forEach((storageType) => {
       describe('getProvider()', () => {
         it('should get file buffer from the storage', async () => {
           const providerPath = `${new Date().getTime()}/provider.tar.gz`;
-          const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'providers', providerPath);
-          const parsedPath = parse(pathToStore);
-          await mkdirp(parsedPath.dir);
-          await writeFile(pathToStore, providerBuf);
+          if (storageType === 'file') {
+            const pathToStore = join(process.env.CITIZEN_STORAGE_PATH, 'providers', providerPath);
+            const parsedPath = parse(pathToStore);
+            await mkdirp(parsedPath.dir);
+            await writeFile(pathToStore, providerBuf);
+          } else if (storageType === 's3') {
+            const buf = Buffer.from('data');
+            s3Mock.on(GetObjectCommand).resolves({
+              Body: Readable.from(buf),
+            });
+          }
 
           const result = await getProvider(providerPath);
           expect(result).to.be.an.instanceof(Buffer);
